pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "deepagents>=0.3.6",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "isort>=7.0.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.16",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-postgres>=3.0.3",
    "langgraph-checkpoint-sqlite>=3.0.2",
    "langsmith>=0.6.3",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tavily-python>=0.7.19",
    "typing>=3.10.0.0",
]
